+++ test/integration/controller_test.rb
@@ -6,4 +6,10 @@ class ControllerTest < IntegrationTest
       assert ActionController::API.method_defined?(:rodauth)
     end
   end
+
+  test "executing controller methods" do
+    visit "/"
+
+    assert_match "controller method", page.text
+  end
 end